REFPLTB-1901 RPI 4 64 Bit Compilation Error in CCSP Misc
build errors in CcspMisc while building RPi4 Broadband 64bit:
| CcspMisc/source/dhcp_client_utils/dhcp_client_common.c:265:35: error: format '%lld' expects argument of type 'long long int', but argument 7 has type 'int64_t' {aka 'long int'} [-Werror=format=]
|   265 |                         DBG_PRINT(%s %d: %s running in pid %lld.. checking for cmdline param %s\n, __FUNCTION__, __LINE__, name, pid, args);
|       |                                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~                                ~~~
|       |                                                                                                                                    |
|       |                                                                                                                                    int64_t {aka long int}
| CcspMisc/source/dhcp_client_utils/dhcp_client_utils.h:64:53: note: in definition of macro 'DBG_PRINT'
|    64 |         RDK_LOG(RDK_LOG_INFO, LOG.RDK.WANMANAGER, fmt, ##arg);
|       |                                                     ^~~
| CcspMisc/source/dhcp_client_utils/dhcp_client_common.c:265:64: note: format string is defined here
|   265 |                         DBG_PRINT(%s %d: %s running in pid %lld.. checking for cmdline param %s\n, __FUNCTION__, __LINE__, name, pid, args);
|       |                                                             ~~~^
|       |                                                                |
|       |                                                                long long int
|       |                                                             %ld
| CcspMisc/source/dhcp_client_utils/dhcp_client_common.c:266:72: error: format '%lld' expects argument of type 'long long int', but argument 4 has type 'int64_t' {aka 'long int'} [-Werror=format=]
|   266 |                         snprintf(filename, sizeof(filename), /proc/%lld/cmdline, pid);
|       |                                                                     ~~~^           ~~~
|       |                                                                        |           |
|       |                                                                        |           int64_t {aka long int}
|       |                                                                        long long int
|       |                                                                     %ld

+++ source/dhcp_client_utils/dhcp_client_common.c
@@ -235,7 +235,7 @@ static int check_proc_entry_for_pid (char * name, char * args)
         if ((dent->d_type == DT_DIR) &&
                 (SUCCESS == strtol64(dent->d_name, NULL, 10, &pid)))
         {
-            snprintf(filename, sizeof(filename), "/proc/%lld/stat", pid);
+            snprintf(filename, sizeof(filename), "/proc/%lld/stat", (long long int) pid);
             fp = fopen(filename, "r");
             if (fp == NULL)
             {
@@ -262,8 +262,8 @@ static int check_proc_entry_for_pid (char * name, char * args)
                     if (args != NULL)
                     {
                         // argument to be verified before returning pid
-                        DBG_PRINT("%s %d: %s running in pid %lld.. checking for cmdline param %s\n", __FUNCTION__, __LINE__, name, pid, args);
-                        snprintf(filename, sizeof(filename), "/proc/%lld/cmdline", pid);
+                        DBG_PRINT("%s %d: %s running in pid %lld.. checking for cmdline param %s\n", __FUNCTION__, __LINE__, name, (long long int) pid, args);
+                        snprintf(filename, sizeof(filename), "/proc/%lld/cmdline", (long long int) pid);
                         fp = fopen(filename, "r");
                         if (fp == NULL)
                         {
